More progress with "generic" image containers.  Implemented iterators on float stream containers...

+++ mpicbg/image/FloatWritable.java
@@ -1,5 +1,18 @@
 package mpicbg.image;
 
+/**
+ * Implements Operator and Writable for float without the actual write and
+ * operation methods for float.  This is necessary to overcome the
+ * limitations of the Java programming language like multiple inheritance
+ * and generics for basic types which otherwise require to write the exact
+ * same code ever and ever again.
+ * 
+ * All these methods tend to be slow because each value has to be casted to
+ * float.  Take care to use compatible types.
+ * 
+ * @author Saalfeld and Preibisch
+ *
+ */
 public abstract class FloatWritable extends Cursor implements Operator, Writable
 {
 	final protected float[] a;
@@ -16,23 +29,23 @@ public abstract class FloatWritable extends Cursor implements Operator, Writable
 	public void add( long c ) {	add( ( float )c ); }
 	public void add( double c ) { add( ( float )c );	}
 
-	public void sub( byte c ){ sub( ( byte )c ); }
-	public void sub( short c ) { sub( ( byte )c ); }
-	public void sub( int c ) { sub( ( byte )c ); }
-	public void sub( long c ) {	sub( ( byte )c ); }
-	public void sub( double c ) { sub( ( byte )c );	}
+	public void sub( byte c ){ sub( ( float )c ); }
+	public void sub( short c ) { sub( ( float )c ); }
+	public void sub( int c ) { sub( ( float )c ); }
+	public void sub( long c ) {	sub( ( float )c ); }
+	public void sub( double c ) { sub( ( float )c );	}
 
-	public void mul( byte c ){ mul( ( byte )c ); }
-	public void mul( short c ) { mul( ( byte )c ); }
-	public void mul( int c ) { mul( ( byte )c ); }
-	public void mul( long c ) {	mul( ( byte )c ); }
-	public void mul( double c ) { mul( ( byte )c );	}
+	public void mul( byte c ){ mul( ( float )c ); }
+	public void mul( short c ) { mul( ( float )c ); }
+	public void mul( int c ) { mul( ( float )c ); }
+	public void mul( long c ) {	mul( ( float )c ); }
+	public void mul( double c ) { mul( ( float )c );	}
 
-	public void div( byte c ){ div( ( byte )c ); }
-	public void div( short c ) { div( ( byte )c ); }
-	public void div( int c ) { div( ( byte )c ); }
-	public void div( long c ) {	div( ( byte )c ); }
-	public void div( double c ) { div( ( byte )c );	}
+	public void div( byte c ){ div( ( float )c ); }
+	public void div( short c ) { div( ( float )c ); }
+	public void div( int c ) { div( ( float )c ); }
+	public void div( long c ) {	div( ( float )c ); }
+	public void div( double c ) { div( ( float )c );	}
 	
 	public void set( Object[] a )
 	{
